+++ unstructured/ingest/processor.py
@@ -77,15 +77,16 @@ class Processor:
         # Debugging tip: use the below line and comment out the mp.Pool loop
         # block to remain in single process
         # self.doc_processor_fn(docs[0])
-
-        with mp.Pool(
-            processes=self.num_processes,
-            initializer=ingest_log_streaming_init,
-            initargs=(logging.DEBUG if self.verbose else logging.INFO,),
-        ) as pool:
-            pool.map(self.doc_processor_fn, docs)
-
-        self.cleanup()
+        logger.info(f"Processing {len(docs)} docs")
+        try:
+            with mp.Pool(
+                processes=self.num_processes,
+                initializer=ingest_log_streaming_init,
+                initargs=(logging.DEBUG if self.verbose else logging.INFO,),
+            ) as pool:
+                pool.map(self.doc_processor_fn, docs)
+        finally:
+            self.cleanup()
 
 
 def process_documents(